# server/openapi.yaml
openapi: 3.0.0
info:
  title: Neko API Reference
  version: "3.0.0"
  description: |
    Neko uses WebRTC and WebSocket besides REST API to provide real-time video streaming, interactive collaboration, and remote desktop functionality.

    ### Core Technologies

    - **Video Streaming over WebRTC**
      - Neko leverages **WebRTC** for real-time, low-latency, high-quality video and audio streaming.
      - Keyboard and mouse events are transmitted over the same WebRTC connection for seamless interaction.

    - **Real-Time Signaling and Communication over WebSocket**
      - **WebSocket** is used for sSignaling in WebRTC (e.g., session establishment, ICE candidate exchange).  
      - Live chat, session updates, and keyboard/mouse fallback are handled over WebSocket.

    - **REST API for Everything Else**
      - **Authentication & Session Management** - Secure user authentication and session handling.  
      - **Room and User Management** - Creating, modifying, and controlling user access to rooms.  
      - **Desktop Controls** - Adjusting screen configurations, managing keyboard and mouse events.
      - **System Health & Metrics** - Retrieving system health checks and operational insights.  

    ### How It All Works Together

    1. **WebRTC** handles media streaming for **real-time video, audio and data**.
    2. **WebSocket** ensures a **persistent, bidirectional** connection for low-latency interactions.  
...

...
  license:
    name: Apache 2.0
...
servers:
...
paths:
...
  /api/stats:
    get:
      tags:
        - general
      summary: Get Stats
      description: Retrieve statistics about the server and user sessions.
      operationId: stats
      responses:
        '200':
          description: Statistics retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  #
  # current session
  #
...
  /api/profile:
    post:
      tags:
        - current-session
      summary: Update Profile
      description: Update the current user's profile without syncing it with the member profile (experimental).
      operationId: profile
      responses:
        '204':
          description: Profile updated successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberProfile'
        required: true

  #
  # sessions
  #
  /api/sessions:
    get:
      tags:
        - sessions
      summary: List Sessions
      description: Retrieve a list of all active sessions.
      operationId: sessionsGet
      responses:
        '200':
          description: Sessions retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionData'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/room/settings:
    get:
      tags:
        - room-settings
      summary: Get Room Settings
      description: Retrieve the current settings of the room.
      operationId: settingsGet
      responses:
        '200':
          description: Room settings retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - room-settings
      summary: Update Room Settings
      description: Update the settings of the room.
      operationId: settingsSet
      responses:
        '204':
          description: Room settings updated successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
        required: true
...
  /api/room/control:
    get:
      tags:
        - room-control
      summary: Get Control Status
      description: Retrieve the current control status of the room.
      operationId: controlStatus
      responses:
        '200':
          description: Control status retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControlStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/room/control/request:
    post:
      tags:
        - room-control
      summary: Request Control
      description: Request control of the room.
      operationId: controlRequest
      responses:
        '204':
          description: Control requested successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: There is already a host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/room/control/release:
    post:
      tags:
        - room-control
      summary: Release Control
      description: Release control of the room.
      operationId: controlRelease
      responses:
        '204':
          description: Control released successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: There is already a host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/room/control/take:
    post:
      tags:
        - room-control
      summary: Take Control
      description: Take control of the room.
      operationId: controlTake
      responses:
        '204':
          description: Control taken successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/room/control/give/{sessionId}:
    post:
      tags:
        - room-control
      summary: Give Control
      description: Give control of the room to a specific session.
      operationId: controlGive
      parameters:
        - in: path
          name: sessionId
          description: The identifier of the session.
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Control given successfully.
        '400':
          description: Target session is not allowed to host.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/room/control/reset:
    post:
      tags:
        - room-control
      summary: Reset Control
      description: Reset the control status of the room.
      operationId: controlReset
      responses:
        '204':
          description: Control reset successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: NEKO_SESSION
      description: Authentication using a session cookie.
    BearerAuth:
      type: http
      scheme: bearer
      description: Authentication using a Bearer token.
    TokenAuth:
      type: apiKey
      in: query
      name: token
      description: Authentication using a token passed as a query parameter.
  responses:
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Unauthorized:
      description: The request requires user authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          description: Detailed error message.
...
    Stats:
      type: object
      properties:
        has_host:
          type: boolean
          description: Indicates if there is a host currently.
        host_id:
          type: string
          description: The ID of the current host, if any.
        server_started_at:
          type: string
          format: date-time
          description: The timestamp when the server started.
        total_users:
          type: integer
          description: The total number of users connected.
        last_user_left_at:
          type: string
          format: date-time
          description: The timestamp when the last user left, if any.
        total_admins:
          type: integer
          description: The total number of admins connected.
        last_admin_left_at:
          type: string
          format: date-time
          description: The timestamp when the last admin left, if any.

    #
    # sessions
    #
...
    SessionData:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the session.
        profile:
          $ref: '#/components/schemas/MemberProfile'
          description: The profile information of the user.
        state:
          $ref: '#/components/schemas/SessionState'
          description: The current state of the session.
    SessionState:
      type: object
      properties:
        is_connected:
          type: boolean
          description: Indicates if the user is connected.
        is_watching:
          type: boolean
          description: Indicates if the user is watching.

    #
    # room
    #
    Settings:
      type: object
      properties:
        private_mode:
          type: boolean
          description: Indicates if the room is in private mode.
        locked_controls:
          type: boolean
          description: Indicates if the room controls are locked.
        implicit_hosting:
          type: boolean
          description: Indicates if implicit hosting is enabled.
        inactive_cursors:
          type: boolean
          description: Indicates if inactive cursors are shown.
        merciful_reconnect:
          type: boolean
          description: Indicates if merciful reconnect is enabled.
        plugins:
          type: object
          additionalProperties: true
          description: Additional plugin settings.
...
    ControlStatus:
      type: object
      properties:
        has_host:
          type: boolean
          description: Indicates if there is a host currently.
        host_id:
          type: string
          description: The ID of the current host, if any.
...
    MemberProfile:
      x-tags:
        - members
      type: object
      properties:
        name:
          type: string
          description: The name of the member.
        is_view_only:
          type: boolean
          description: Indicates that the member is passive and all interactive capabilities are denied server-side.
        is_admin:
          type: boolean
          description: Indicates if the member is an admin.
        can_login:
          type: boolean
          description: Indicates if the member can log in.
        can_connect:
          type: boolean
          description: Indicates if the member can connect.
        can_watch:
          type: boolean
          description: Indicates if the member can watch.
        can_host:
          type: boolean
          description: Indicates if the member can host.
        can_share_media:
          type: boolean
          description: Indicates if the member can share media.
        can_access_clipboard:
          type: boolean
          description: Indicates if the member can access the clipboard.
        sends_inactive_cursor:
          type: boolean
          description: Indicates if the member sends inactive cursor.
        can_see_inactive_cursors:
          type: boolean
          description: Indicates if the member can see inactive cursors.
        plugins:
          type: object
          additionalProperties: true
          description: Additional plugin settings.
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - TokenAuth: []